pyproject.toml
[project]
name = "pydiscogsqrcodegenerator"
version = "0.1.0"
description = "Generate QR code stickers for your Discogs record collection"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "flask>=3.0",
    "flask-session>=0.8",
    "flask-sqlalchemy>=3.1",
    "python3-discogs-client>=2.7",
    "python-dotenv>=1.0",
    "fpdf2>=2.8",
    "segno>=1.6",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pydiscogsqrcodegenerator"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
